Add test for fas home page
